implement rate limiting

+++ src/server/routers/_app.ts
@@ -1,9 +1,17 @@
 import { Label, Status } from "@prisma/client";
 import { prisma } from "@src/lib/prisma";
 import deleteFile from "@src/lib/upload";
+import { Ratelimit } from "@upstash/ratelimit";
+import { Redis } from "@upstash/redis";
 import { Configuration, OpenAIApi } from "openai";
 import { z } from "zod";
 import { protectedProcedure, publicProcedure, router } from "../trpc";
+import { TRPCError } from "@trpc/server";
+
+const ratelimiter = new Ratelimit({
+  redis: Redis.fromEnv(),
+  limiter: Ratelimit.slidingWindow(3, "5 s"),
+});
 
 const configuration = new Configuration({
   apiKey: process.env.OPENAI_API_KEY,
@@ -37,6 +45,10 @@ export const appRouter = router({
   getProfile: protectedProcedure
     .input(z.string().optional())
     .query(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.profile.findUnique({
         where: {
           id: input ?? ctx.auth.userId,
@@ -44,6 +56,10 @@ export const appRouter = router({
       });
     }),
   getInbox: protectedProcedure.query(async ({ ctx }) => {
+    const { success } = await ratelimiter.limit(ctx.auth.userId);
+    if (!success) {
+      throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+    }
     const [sending, recieving] = await prisma.$transaction([
       prisma.post.findMany({
         where: {
@@ -87,7 +103,11 @@ export const appRouter = router({
       return 0;
     });
   }),
-  getRanking: protectedProcedure.query(async () => {
+  getRanking: protectedProcedure.query(async ({ ctx }) => {
+    const { success } = await ratelimiter.limit(ctx.auth.userId);
+    if (!success) {
+      throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+    }
     return await prisma.post.findMany({
       where: {
         label: "PUBLIC",
@@ -123,6 +143,10 @@ export const appRouter = router({
     });
   }),
   getBanners: protectedProcedure.query(async ({ ctx }) => {
+    const { success } = await ratelimiter.limit(ctx.auth.userId);
+    if (!success) {
+      throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+    }
     return await prisma.banner.findMany({
       orderBy: {
         id: "asc",
@@ -137,6 +161,10 @@ export const appRouter = router({
     });
   }),
   getProfileBanners: protectedProcedure.query(async ({ ctx }) => {
+    const { success } = await ratelimiter.limit(ctx.auth.userId);
+    if (!success) {
+      throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+    }
     return await prisma.banner.findMany({
       where: {
         purchases: {
@@ -172,6 +200,10 @@ export const appRouter = router({
     });
   }),
   getProfilePosts: protectedProcedure.query(async ({ ctx }) => {
+    const { success } = await ratelimiter.limit(ctx.auth.userId);
+    if (!success) {
+      throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+    }
     return await prisma.post.findMany({
       where: {
         authorId: ctx.auth.userId,
@@ -193,6 +225,10 @@ export const appRouter = router({
     });
   }),
   getBookmarkedPosts: protectedProcedure.query(async ({ ctx }) => {
+    const { success } = await ratelimiter.limit(ctx.auth.userId);
+    if (!success) {
+      throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+    }
     return await prisma.post.findMany({
       where: {
         bookmarks: {
@@ -218,6 +254,10 @@ export const appRouter = router({
     });
   }),
   getFriends: protectedProcedure.query(async ({ ctx }) => {
+    const { success } = await ratelimiter.limit(ctx.auth.userId);
+    if (!success) {
+      throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+    }
     const [sending, recieving] = await prisma.$transaction([
       prisma.friend.findMany({
         where: {
@@ -252,6 +292,10 @@ export const appRouter = router({
     });
   }),
   getAllFriends: protectedProcedure.query(async ({ ctx }) => {
+    const { success } = await ratelimiter.limit(ctx.auth.userId);
+    if (!success) {
+      throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+    }
     const [sending, recieving] = await prisma.$transaction([
       prisma.friend.findMany({
         where: {
@@ -297,6 +341,10 @@ export const appRouter = router({
   getSendingFriendStatus: protectedProcedure
     .input(String)
     .query(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.friend.findUnique({
         where: {
           receiverId_senderId: {
@@ -309,6 +357,10 @@ export const appRouter = router({
   getRecievingFriendStatus: protectedProcedure
     .input(String)
     .query(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.friend.findUnique({
         where: {
           receiverId_senderId: {
@@ -325,7 +377,11 @@ export const appRouter = router({
         recieverId: z.string(),
       })
     )
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.friend.create({
         data: {
           senderId: input.senderId,
@@ -341,7 +397,11 @@ export const appRouter = router({
         status: z.nativeEnum(Status),
       })
     )
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.friend.update({
         data: {
           status: input.status,
@@ -362,7 +422,11 @@ export const appRouter = router({
         recieverId: z.string(),
       })
     )
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.friend.delete({
         where: {
           receiverId_senderId: {
@@ -386,6 +450,10 @@ export const appRouter = router({
       })
     )
     .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return input.friendId
         ? await prisma.$transaction([
             prisma.post.create({
@@ -453,7 +521,11 @@ export const appRouter = router({
         friendId: z.string().optional(),
       })
     )
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return input.friendId
         ? await prisma.post.update({
             data: {
@@ -491,7 +563,11 @@ export const appRouter = router({
         id: z.number(),
       })
     )
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.post.delete({
         where: {
           id: input.id,
@@ -501,6 +577,10 @@ export const appRouter = router({
   getLikes: protectedProcedure
     .input(z.string().optional())
     .query(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.post.findMany({
         where: {
           authorId: input ?? ctx.auth.userId,
@@ -521,7 +601,11 @@ export const appRouter = router({
         profileId: z.string(),
       })
     )
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.like.create({
         data: {
           postId: input.postId,
@@ -536,7 +620,11 @@ export const appRouter = router({
         profileId: z.string(),
       })
     )
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.like.delete({
         where: {
           postId_profileId: {
@@ -549,6 +637,10 @@ export const appRouter = router({
   getBookmarks: protectedProcedure
     .input(z.string().optional())
     .query(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.post.findMany({
         where: {
           authorId: input ?? ctx.auth.userId,
@@ -569,7 +661,11 @@ export const appRouter = router({
         profileId: z.string(),
       })
     )
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.bookmark.create({
         data: {
           postId: input.postId,
@@ -584,7 +680,11 @@ export const appRouter = router({
         profileId: z.string(),
       })
     )
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.bookmark.delete({
         where: {
           postId_profileId: {
@@ -597,6 +697,10 @@ export const appRouter = router({
   getComments: protectedProcedure
     .input(z.string().optional())
     .query(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.post.findMany({
         where: {
           authorId: input ?? ctx.auth.userId,
@@ -627,6 +731,10 @@ export const appRouter = router({
       })
     )
     .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.comment.create({
         data: {
           postId: input.postId,
@@ -641,7 +749,11 @@ export const appRouter = router({
         id: z.number(),
       })
     )
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.comment.delete({
         where: {
           id: input.id,
@@ -656,7 +768,11 @@ export const appRouter = router({
         credits: z.number(),
       })
     )
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.$transaction([
         prisma.purchase.create({
           data: {
@@ -681,6 +797,10 @@ export const appRouter = router({
       })
     )
     .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await prisma.profile.update({
         data: {
           banner: input.banner,
@@ -696,7 +816,11 @@ export const appRouter = router({
         attachmentPath: z.string(),
       })
     )
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       return await deleteFile({
         accountId: process.env.UPLOAD_ACCOUNTID,
         apiKey: process.env.UPLOAD_SECRETKEY,
@@ -706,6 +830,10 @@ export const appRouter = router({
       });
     }),
   deleteProfile: protectedProcedure.mutation(async ({ ctx }) => {
+    const { success } = await ratelimiter.limit(ctx.auth.userId);
+    if (!success) {
+      throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+    }
     return await fetch("https://api.clerk.com/v1/users/" + ctx.auth.userId, {
       method: "DELETE",
       headers: {
@@ -721,6 +849,10 @@ export const appRouter = router({
       })
     )
     .mutation(async ({ input, ctx }) => {
+      const { success } = await ratelimiter.limit(ctx.auth.userId);
+      if (!success) {
+        throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+      }
       const completion = await openai.createCompletion({
         model: "text-davinci-003",
         prompt:
